requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
requests
yfinance
flask_table
googlesearch-python
gunicorn